style: move getter/setter location
