+++ src/index.umd.ts
@@ -9,16 +9,11 @@ import * as Control from "./control";
 import * as Renderer from "./renderer";
 import * as Constants from "./const/external";
 import { merge } from "./utils";
-// import { DEFAULT_OPTIONS, MOVE_TYPE } from "./consts";
 
 merge(Flicking, Core);
 merge(Flicking, Camera);
 merge(Flicking, Control);
 merge(Flicking, Renderer);
 merge(Flicking, Constants);
-
-// (Flicking as any).withFlickingMethods = withFlickingMethods;
-// (Flicking as any).DEFAULT_OPTIONS = DEFAULT_OPTIONS;
-// (Flicking as any).MOVE_TYPE = MOVE_TYPE;
 
 export default Flicking;
