+++ src/renderer/index.ts
@@ -2,13 +2,14 @@
  * Copyright (c) 2015 NAVER Corp.
  * egjs projects are licensed under the MIT license
  */
-import Renderer from "./Renderer";
+import Renderer, { RendererOptions } from "./Renderer";
 import RawRenderer from "./RawRenderer";
 import VisibleRenderer from "./VisibleRenderer";
 import ExternalRenderer from "./ExternalRenderer";
 
 export {
   Renderer,
+  RendererOptions,
   RawRenderer,
   VisibleRenderer,
   ExternalRenderer
